add support to make container a promise or value during config . Lets me defer for containers which might not have rendered yet inside the app
