+++ demo/src/app/app.component.html
@@ -79,3 +79,4 @@
     </div>
   </div>
 </div>
+<div class="modal-container"></div>
