+++ src/dialog-options.ts
@@ -1,0 +1,6 @@
+export interface DialogOptions {
+  index?: number;
+  autoCloseTimeout?: number;
+  closeByClickingOutside?: boolean;
+  backdropColor?: string;
+}
