+++ src/index.ts
@@ -1,3 +1,4 @@
 export { DialogModalModule } from './dialog-modal.module';
 export { DialogComponent } from './dialog.component';
-export { DialogService, DialogOptions, DialogServiceConfig } from './dialog.service';
+export { DialogOptions }  from './dialog-options';
+export { DialogService, DialogServiceConfig } from './dialog.service';
